Move store settings to a separate admin folder
