+++ react/components/admin/institutional/Form/index.tsx
@@ -5,16 +5,14 @@ import { defineMessages, InjectedIntlProps, injectIntl } from 'react-intl'
 
 import { ToastConsumerFunctions } from 'vtex.styleguide'
 
+import { formatStatements } from '../../../../utils/conditions'
+import { FormErrors } from '../../pages/Form/typings'
+import { generateNewRouteId } from '../../pages/Form/utils'
+import { FormProps } from '../../store/StoreForm/components/withStoreSettings'
+
 import Form from './Form'
 import { getValidateFormState } from './utils'
 import { OperationsResults } from './withContentContext'
-
-import { FormErrors } from '../../pages/Form/typings'
-import { generateNewRouteId } from '../../pages/Form/utils'
-
-import { FormProps } from '../../../EditorContainer/StoreEditor/Store/StoreForm/components/withStoreSettings'
-
-import { formatStatements } from '../../../../utils/conditions'
 
 interface ComponentProps {
   culture: RenderContext['culture']
