+++ react/components/admin/store/PWAForm/mutations/UpdateManifest.graphql
