+++ react/components/admin/store/PWAForm/mutations/UpdateManifestIcon.graphql
