+++ react/components/admin/store/PWAForm/mutations/UpdatePWASettings.graphql
