+++ react/components/admin/store/StoreForm/index.tsx
@@ -7,10 +7,10 @@ import Form, { UiSchema } from 'react-jsonschema-form'
 import { formatIOMessage } from 'vtex.native-types'
 import { Button, ToastContext } from 'vtex.styleguide'
 
-import ArrayFieldTemplate from '../../../../form/ArrayFieldTemplate'
-import BaseInput from '../../../../form/BaseInput'
-import FieldTemplate from '../../../../form/FieldTemplate'
-import ObjectFieldTemplate from '../../../../form/ObjectFieldTemplate'
+import ArrayFieldTemplate from '../../../form/ArrayFieldTemplate'
+import BaseInput from '../../../form/BaseInput'
+import FieldTemplate from '../../../form/FieldTemplate'
+import ObjectFieldTemplate from '../../../form/ObjectFieldTemplate'
 
 import withStoreSettings, { FormProps } from './components/withStoreSettings'
 import SaveAppSettings from './mutations/SaveAppSettings.graphql'
