+++ react/components/admin/store/StoreForm/mutations/SaveAppSettings.graphql
