+++ react/components/admin/store/StoreForm/queries/AvailableApp.graphql
