+++ react/components/admin/store/StoreForm/queries/InstalledApp.graphql
